fix: require path error due to refactor
