+++ package.json
@@ -2,7 +2,7 @@
   "name": "clear-cra",
   "version": "1.1.22",
   "description": "Clears boilerplate content by deleting unnecessary files and changes code to a required minimum.",
-  "main": "index.js",
+  "main": "bin/index.js",
   "scripts": {
     "prepublish": "npm run format",
     "format": "prettier --write ./**/*.{js,json}",
@@ -26,7 +26,7 @@
   },
   "homepage": "https://github.com/bqardi/clear-cra#readme",
   "bin": {
-    "clear-cra": "index.js"
+    "clear-cra": "bin/index.js"
   },
   "dependencies": {
     "chalk": "^4.1.2",
